+++ pyccel/codegen/printing/fcode.py
@@ -31,7 +31,7 @@ from pyccel.ast.core import FunctionAddress
 from pyccel.ast.core import Return, Module, For
 from pyccel.ast.core import Import, CodeBlock, AsName, EmptyNode
 from pyccel.ast.core import Assign, AliasAssign, Declare, Deallocate
-from pyccel.ast.core import FunctionCall, DottedFunctionCall, PyccelFunctionDef
+from pyccel.ast.core import FunctionCall, PyccelFunctionDef
 
 from pyccel.ast.datatypes import NativeSymbol, NativeString, str_dtype
 from pyccel.ast.datatypes import NativeInteger, NativeBool, NativeFloat, NativeComplex
@@ -176,7 +176,7 @@ type_to_print_format = {
         ('complex'): '"(",F0.12," + ",F0.12,")"',
         ('int'): 'I0',
         ('bool'): 'A',
-        ('string'): 'A',
+        ('str'): 'A',
         ('tuple'):  '*'
 }
 
@@ -437,6 +437,36 @@ class FCodePrinter(CodePrinter):
 
         return decs
 
+    def _calculate_class_names(self, expr):
+        """
+        Calculate the class names of the functions in a class.
+
+        Calculate the names that will be referenced from the class
+        for each function in a class. Also rename magic methods.
+
+        Parameters
+        ----------
+        expr : ClassDef
+            The class whose functions should be renamed.
+        """
+        scope = expr.scope
+        name = expr.name
+        for method in expr.methods:
+            if not method.is_inline:
+                m_name = method.name
+                if m_name in _default_methods:
+                    suggested_name = _default_methods[m_name]
+                    scope.rename_function(method, suggested_name)
+                method.cls_name = scope.get_new_name(f'{name}_{method.name}')
+        for i in expr.interfaces:
+            for f in i.functions:
+                if not f.is_inline:
+                    i_name = f.name
+                    if i_name in _default_methods:
+                        suggested_name = _default_methods[i_name]
+                        scope.rename_function(f, suggested_name)
+                    f.cls_name = scope.get_new_name(f'{name}_{f.name}')
+
     # ============ Elements ============ #
     def _print_PyccelSymbol(self, expr):
         return expr
@@ -449,14 +479,16 @@ class FCodePrinter(CodePrinter):
             name = self._print(expr.name)
         name = name.replace('.', '_')
         if not name.startswith('mod_') and self.prefix_module:
-            name = '{prefix}_{name}'.format(prefix=self.prefix_module,
-                                            name=name)
+            name = f'{self.prefix_module}_{name}'
 
         imports = ''.join(self._print(i) for i in expr.imports)
 
         # Define declarations
         decs = ''
         # ...
+        for c in expr.classes:
+            self._calculate_class_names(c)
+
         class_decs_and_methods = [self._print(i) for i in expr.classes]
         if not isinstance(expr, BindCModule):
             decs += '\n'.join(c[0] for c in class_decs_and_methods)
@@ -1726,8 +1758,8 @@ class FCodePrinter(CodePrinter):
 
         if isinstance(var.dtype, CustomDataType):
             Pyccel__del = expr.variable.cls_base.scope.find('__del__')
-            Pyccel_del_args = [FunctionCallArgument(arg) for arg in Pyccel__del.arguments]
-            return self._print(DottedFunctionCall(Pyccel__del, Pyccel_del_args, var))
+            Pyccel_del_args = [FunctionCallArgument(var)]
+            return self._print(FunctionCall(Pyccel__del, Pyccel_del_args))
 
         if var.is_alias:
             return ''
@@ -1955,20 +1987,7 @@ class FCodePrinter(CodePrinter):
         self.set_scope(expr.scope)
 
 
-        name = self._print(expr.name)
-
-        if expr.cls_name:
-            for k, m in list(_default_methods.items()):
-                name = name.replace(k, m)
-
-            cls_name = expr.cls_name
-            if not (cls_name == '__UNDEFINED__'):
-                name = '{0}_{1}'.format(cls_name, name)
-        else:
-            for i in _default_methods:
-                # because we may have a class Point with init: Point___init__
-                if i in name:
-                    name = name.replace(i, _default_methods[i])
+        name = expr.cls_name or expr.name
 
         sig_parts = self.function_signature(expr, name)
         bind_c = ' bind(c)' if isinstance(expr, BindCFunctionDef) else ''
@@ -2047,17 +2066,12 @@ class FCodePrinter(CodePrinter):
 
         aliases = []
         names   = []
-        ls = [self._print(i.name) for i in expr.methods if not i.is_inline]
-        for i in ls:
-            j = _default_methods.get(i,i)
-            aliases.append(j)
-            names.append('{0}_{1}'.format(name, self._print(j)))
-        methods = ''.join('procedure :: {0} => {1}\n'.format(i, j) for i, j in zip(aliases, names))
+        methods = ''.join(f'procedure :: {method.name} => {method.cls_name}\n' for method in expr.methods)
         for i in expr.interfaces:
-            names = ','.join('{0}_{1}'.format(name, self._print(j.name)) for j in i.functions if not j.is_inline)
+            names = ','.join(f.cls_name for f in i.functions if not f.is_inline)
             if names:
-                methods += 'generic, public :: {0} => {1}\n'.format(self._print(i.name), names)
-                methods += 'procedure :: {0}\n'.format(names)
+                methods += f'generic, public :: {i.name} => {names}\n'
+                methods += f'procedure :: {names}\n'
 
 
 
@@ -2953,8 +2967,19 @@ class FCodePrinter(CodePrinter):
         parent_assign = expr.get_direct_user_nodes(lambda x: isinstance(x, Assign))
         is_function =  len(func_results) == 1 and func_results[0].rank == 0
 
-        if isinstance(expr, DottedFunctionCall):
+        if func.arguments and func.arguments[0].bound_argument:
+            class_variable = args[0].value
             args = args[1:]
+            if isinstance(class_variable, FunctionCall):
+                base = class_variable.funcdef.results[0].var
+                if (not self._additional_code):
+                    self._additional_code = ''
+                var = self.scope.get_temporary_variable(base)
+
+                self._additional_code = self._additional_code + self._print(Assign(var, class_variable)) + '\n'
+                f_name = f'{self._print(var)} % {f_name}'
+            else:
+                f_name = f'{self._print(class_variable)} % {f_name}'
 
         if (not self._additional_code):
             self._additional_code = ''
@@ -3027,24 +3052,6 @@ class FCodePrinter(CodePrinter):
             return '{0} = {1}\n'.format(self._print(results[0]), code)
         else:
             return code
-
-#=======================================================================================
-
-    def _print_DottedFunctionCall(self, expr):
-        if isinstance(expr.prefix, FunctionCall):
-            base = expr.prefix.funcdef.results[0].var
-            if (not self._additional_code):
-                self._additional_code = ''
-            var_name = self.scope.get_new_name()
-            var = base.clone(var_name)
-
-            self.scope.insert_variable(var)
-
-            self._additional_code = self._additional_code + self._print(Assign(var,expr.prefix)) + '\n'
-            user_node = expr.current_user_node
-            expr = DottedFunctionCall(expr.funcdef, expr.args, var)
-            expr.set_current_user_node(user_node)
-        return self._print_FunctionCall(expr)
 
 #=======================================================================================